Move cached_revision from downward lab to lab
